Move testcase to workaround import side effects
