+++ test/dynamo/test_functions.py
@@ -11,6 +11,7 @@ import torch
 from torch import sub
 from torch.nn import functional as F
 
+import torchdynamo.test_case
 import torchdynamo.testing
 from torchdynamo.testing import requires_static_shapes
 
@@ -53,7 +54,7 @@ def inline_unused(x):
     return x + 5.6
 
 
-class FunctionTests(torchdynamo.testing.TestCase):
+class FunctionTests(torchdynamo.test_case.TestCase):
     @make_test
     def test_inline_jit_annotations(x):
         x = inline_script_if_tracing(x)
@@ -670,6 +671,6 @@ class FunctionTests(torchdynamo.testing.TestCase):
 
 
 if __name__ == "__main__":
-    from torchdynamo.testing import run_tests
+    from torchdynamo.test_case import run_tests
 
     run_tests()
