+++ test/dynamo/test_optimizations.py
@@ -8,6 +8,7 @@ from unittest.mock import patch
 import torch
 
 import torchdynamo
+import torchdynamo.test_case
 from torchdynamo.optimizations import backends
 from torchdynamo.optimizations.analysis import has_mutation
 from torchdynamo.optimizations.log_args import conv_args_analysis
@@ -65,7 +66,7 @@ class Conv_Bn_Relu(torch.nn.Module):
         return self.relu(self.bn(self.conv(x)))
 
 
-class TestOptimizations(torchdynamo.testing.TestCase):
+class TestOptimizations(torchdynamo.test_case.TestCase):
     def test_inplacifier(self):
         gm = torch.fx.symbolic_trace(Seq())
         normalize(gm)
@@ -184,7 +185,7 @@ class TestOptimizations(torchdynamo.testing.TestCase):
         self.assertEqual(r2.dtype, torch.bfloat16)
 
 
-class NormalizeIRTests(torchdynamo.testing.TestCase):
+class NormalizeIRTests(torchdynamo.test_case.TestCase):
     @unittest.skipIf(not has_functorch(), "requires functorch")
     def test_inplace_normalize(self):
         def fn(a, b):
@@ -203,6 +204,6 @@ class NormalizeIRTests(torchdynamo.testing.TestCase):
 
 
 if __name__ == "__main__":
-    from torchdynamo.testing import run_tests
+    from torchdynamo.test_case import run_tests
 
     run_tests()
