+++ test/dynamo/test_skip_non_tensor.py
@@ -4,10 +4,11 @@ from unittest.mock import patch
 import torch
 
 import torchdynamo
+import torchdynamo.test_case
 from torchdynamo.testing import CompileCounter
 
 
-class SkipNonTensorTests(torchdynamo.testing.TestCase):
+class SkipNonTensorTests(torchdynamo.test_case.TestCase):
     def test_add_tensor1(self):
         def fn(a, b):
             return a + b
@@ -107,6 +108,6 @@ class SkipNonTensorTests(torchdynamo.testing.TestCase):
 
 
 if __name__ == "__main__":
-    from torchdynamo.testing import run_tests
+    from torchdynamo.test_case import run_tests
 
     run_tests()
