+++ test/dynamo/test_subgraphs.py
@@ -4,6 +4,7 @@ from unittest.mock import patch
 
 import torch
 
+import torchdynamo.test_case
 import torchdynamo.testing
 from torchdynamo import config
 from torchdynamo.testing import unsupported
@@ -17,7 +18,7 @@ def indirectly_unsupported(a, b):
     return unsupported(a, c)
 
 
-class SubGraphTests(torchdynamo.testing.TestCase):
+class SubGraphTests(torchdynamo.test_case.TestCase):
     def _common(self, fn, frame_count, op_count):
         torchdynamo.reset()
         v1 = torch.ones(10)
@@ -528,6 +529,6 @@ class SubGraphTests(torchdynamo.testing.TestCase):
 
 
 if __name__ == "__main__":
-    from torchdynamo.testing import run_tests
+    from torchdynamo.test_case import run_tests
 
     run_tests()
