+++ test/inductor/test_torchinductor.py
@@ -4022,6 +4022,6 @@ if HAS_CUDA:
 
 
 if __name__ == "__main__":
-    from torchdynamo.testing import run_tests
+    from torchdynamo.test_case import run_tests
 
     run_tests(needs="filelock")
